LayoutTests/editing/input/firstrectforcharacterrange-styled-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 1 of #text > I > B > BODY > HTML > #document to 1 of #text > I > B > BODY > HTML > #document toDOMRange:range from 1 of #text > I > B > BODY > HTML > #document to 3 of #text > I > B > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
123456
10,572,0,18
18,572,0,18
58,572,0,18
58,572,0,18
10,572,8,18
50,572,8,18
50,572,8,18
50,572,8,18
10,572,48,18
10,572,48,18
10,572,48,18
18,572,40,18
0,0,0,0
0,0,0,0
0,0,0,0
0,0,0,0
